producer/requirements.txt
fdk
kafka-python